requirements.txt
transformers
gradio
torch
sentencepiece
accelerate
supabase
python-dotenv
